requirements.txt

gspread
twilio
python-binance
google-auth
google-auth-oauthlib
